Fix transient dependencies between generated files
It turns out, if a generated file is an input to another generated file, CMake will not create a chain. It will marely add all the transient dependencies to any target that depends on the final file, but it will not build them in any particular order. This is solved using an auxiliary target accompanying each custom command.