Fix cube map
After a lot of pain! The most important part to remember is that the
textures need to be flipped in the X-direction before using as cube faces.
Tis is because the faces appear normal when viewed from OUTSIDE the
cube, but we are INSIDE. Alternatively, we could just apply the
mirroring already at the rendering stage and incorporate it into the
projection matrix. But this brings with itself potential problems with
culling and other screen-space effects. Rendering with the ordinary
right-handed coordinate system will save a lot problems and flipping the
texture has to be done in one spot and remains stable.